Query with json

+++ api/src/db.py
@@ -2,6 +2,7 @@ import mysql.connector
 import warnings
 from config import config
 import time
+import json
 
 
 class Db:
@@ -24,15 +25,17 @@ class Db:
         self.connection = mysql.connector.connect(**self.config)
         self.cursor = self.connection.cursor()
 
-        if config.clean_database is True:
-            self.cursor.execute(f'DROP DATABASE {self.db_name}')
-            self.connection.commit()
-
         self.cursor.execute('SHOW DATABASES')
         databases = [row[0] for row in self.cursor.fetchall()]
 
+        if config.clean_database is True:
+            if self.db_name in databases:
+                self.cursor.execute(f'DROP DATABASE {self.db_name}')
+                self.connection.commit()
+
         if self.db_name not in databases:
             self.cursor.execute(f'CREATE DATABASE IF NOT EXISTS {self.db_name}')
+            self.connection.commit()
 
         self.cursor.close()
         self.connection.close()
@@ -41,6 +44,7 @@ class Db:
 
         self.connection = mysql.connector.connect(**self.config)
         self.cursor = self.connection.cursor()
+        self.cursor_dict = self.connection.cursor(dictionary=True)
 
         self.cursor.execute('SHOW TABLES')
         tables = [row[0] for row in self.cursor.fetchall()]
@@ -116,6 +120,14 @@ class Db:
         query += f' WHERE wechat_openid={wechat_openid}' if wechat_openid is not None else ''
         self.cursor.execute(query)
         return int(self.cursor.fetchone()[0])
+
+    def get_simulators(self, wechat_openid: str | None, offset: int, limit: int):
+        query = f'SELECT * FROM {self.table_simulators}'
+        query += f' WHERE wechat_openid={wechat_openid}' if wechat_openid is not None else ''
+        query += ' ORDER BY timestamp DESC'
+        query += f' LIMIT {limit} OFFSET {offset}'
+        self.cursor_dict.execute(query)
+        return self.cursor_dict.fetchall()
 
     def ban(self, wechat_openid, ban_by_reason, ban_by_id):
         self.cursor.execute(
